requirements.txt
dask==1.2.2
Django==1.11.22
django-extensions==2.1.6
django-mathfilters==0.4.0
genson==1.1.0
numpy==1.16.3
pandas==0.24.2
PyYAML==5.1.2
requests==2.22.0
SQLAlchemy==1.3.3
